requirements.txt
fastapi>=0.111.0
uvicorn>=0.30.0
pdfplumber>=0.11.0
pydantic>=2.7.0
openai>=1.30.0
python-dotenv>=1.0.1
python-multipart>=0.0.9
